First Commit
zCrypt is a utility program which provides straight-forward, powerful
functions for protect, secure, crypt or split files and folders.

zCrypt is lightweigth, portable, and very easy to use.

zCrypt is compiled with Dotnet Core framework. That's why it works on
various platforms (Windows, Linux and Mac).

zCrypt is developped by http://www.zem.fr team and it's absolutely free
for use.